Suggest a "docker cdff" command
"docker run" commands often involve a large number of options that
give fine-grained control to the operator who starts a container
from an image. This is the case of the "docker run" command used to
create a container from the InFuse Docker images: there can be
options to bind-mount files and directories from the user's computer
into the container, options to share the X and dev-log sockets,
options to use an init process and delete the container on exit, and
so on, to the discretion of the user.

This commit provides a shell function, that the user can copy-paste
into one of their shell startup files, that:

* Defines "docker cdff" as shorthand for a long CDFF-specific
  "docker run" command
* Replaces the "alias docker='sudo /usr/bin/docker'" line that is
  commonly put in ~/.bashrc or ~/.bash_aliases

The function and its documentation are in Tools/Docker/docker_aliases
.template.

The provided functionality is reminiscent of Git aliases which are
custom Git subcommands defined by the user. If and when Docker
implements a similar feature to Git aliases, it may be used to
re-implement the functionality provided by this commit.