Call the Catch executable directly rather than through CTest
Running the unit tests using "make test" (an alias of "ctest") filters out a lot
of useful information from the test executable(s). In particular, it just
reports failure, but doesn't print which test case failed, and why. This commit
fixes that issue by calling the test executable directly during CI, rather than
through CTest.

An alternative would be to call the test executable using "ctest --verbose",
it's pretty much equivalent thanks to the definition of an adequate test target
in a CMake listfile. Another alternative would be to register each Catch test
case as a CTest test: the documentation of Catch explains how to do that, but
I haven't looked into it, as it doesn't seem to be high-priority to me.