Make dependencies quiet during compilation
The dependencies of CDFF-core (OpenCV, PCL, VTK...) output a lot of warnings
during compilation, partly because of the -Wall option. We cannot fix those
warnings, and they hide the warnings coming from our own code.

This commit fixes that issue by declaring the dependencies as system libraries
to the compiler (through CMake). This makes the compiler, or at least GNU g++
because that feature is actually compiler-dependent, ignore warnings coming
from them, just as it ignores warnings comming from the standard libraries
installed in the system.

I suppose this is realized by CMake passing the library headers to the compiler
using the option -isystem instead of -I.