Remove unnecessary ESROCOS Core types
As far as I know we're using the ESROCOS Base types and Sensor types, not the
ESROCOS Core types, which appear to be core types for ESROCOS but not helpful
for InFuse (ApplicationCodes, ComponentLifecycle, Logger, Mars, Vizkit3D).

I haven't found a trace of these types in either CDFF or CDFF_dev, they don't
appear to be dependencies of other types either, and I haven't found a mention
of them in the different tables about data types that have been circulated,
at least not in those that I know about.

Removing them because of clutter.