Port all ASN.1 types from the CTypes:: namespace to the asn1Scc prefix
The compiled ASN.1 types used to be imported in the CDFF in a C++ namespace
called CTypes, for instance: CTypes::Time, CTypes::Frame.

However, it appears that OG1-ESROCOS and TASTE folks compile the ASN.1 types
with an asn1Scc prefix added to the front, for instance: asn1SccTime,
asn1SccFrame. C can't do namespaces, only prefix conventions.

In order to maximize compatibility with ESROCOS, for example when replaying,
in CDFF-dev, logs made using ESROCOS, we are replacing the C++ CTypes::
namespace with the asn1Scc prefix. This commit performs the bulk of this
replacement.

Thankfully, because most of the CDFF doesn't use the compiled ASN.1 types
directly, but through wrappers defined in Common/Types/CPP, the changes are
mostly limited to these wrappers.

Other changes: streamline comments, fix include guards.