Update I/Os for ForceMesh DFN & HapticScanning DFPC
The interface was changed so that the DFN & DFPC take a single measurement as their input and output the complete pointcloud. The previous interface has the DFN processing the entire list of points each time the `process` method was called. 

This change makes it less costly to process clouds incrementally.