Stop looking for Boost.Serialization and Boost.Timer
The CDFF, master branch, as of now, does not depend on these libraries.

Not looking for them does not prevent linking against them in branches
where the CDFF does depend on these libraries. It does prevent using
the imported targets Boost::serialization and Boost::timer though, but
then again, importing those targets is a modification that belongs to
the feature branches where they need to be used, not to master.