Highlight how necessary it is to use target-specific include dirs
There is no initial default value for the directories that are searched
for header files: that list is initially empty. We do not have to worry
about whether to append or prepend the include directories we add to
this list with include_directories().

This can be investigated with:

get_property(includedirs DIRECTORY PROPERTY INCLUDE_DIRECTORIES)
message(STATUS ">>>>>> INCLUDE_DIRECTORIES directory property is now "
  "set to this value in the current directory: ${includedirs}")