Use GNU readlink to find /path/to/CDFF/Tools/Docker
I'm not sure how critical it is to resolve symlinks and ensure that we have
an absolute path to the directory that contains the installer, but if it is,
GNU readlink has the options -f|-e|-m to canonicalize pathnames by following
every symlink in every component of the given pathname recursively until all
symlinks have been resolved, and it is robust to symlink cycles too.

Of course it isn't POSIX but neither is most of this script anyway. BSD and
MacOS systems can use greadlink from a port of GNU coreutils, if I understand
correctly.