Remove unused Image type (original Frame type from original ESROCOS)
Nothing in the CDFF, at least not in the master branch, uses the Image
ASN.1 type anymore, which is the original Frame type from ESROCOS (as
ESROCOS was when that type was copied from it), just renamed.

AFAIK partners have stopped using it around Sep. 2018 at the latest.

This commit removes the ASN.1 type, its C++ wrapper, and a Frame-to-Image
converter from the master branch. If they are still of use in some topic
branch, they can stay in that topic branch.